Managed three cases for cloning repository locally
Added functionality to three cases that appear when locally cloning a
repository. 1. not allowing a clone when there exists a .git folder, 2.
not allowing a clone when a folder is empty, and the third case if there
is a directory with files, it will create a seperate folder inside and
clone the repository there.